requirements.txt
pytest
discord.py==2.0.1
packaging==21.3
python-dotenv==0.20.0
replicate==0.0.1a15